backend/requirements-build.txt
# LabelGen Build Requirements
# Install with: pip install -r requirements-build.txt

# Runtime requirements
django==6.0.2

# PyInstaller for creating .exe (use latest available)
pyinstaller>=6.15.0

# System tray application
pystray>=0.19.0
Pillow>=10.0.0

# Note: Version ranges allow for compatibility with future updates
